Fix search result outline being cut-off
This PR removes an `overflow: hidden;` that caused the focus outline to
be cut-off. No matter if we use `box-shadow` or `outline`, hiding the
overflow would always cut this off.

I tried to find out why the `overflow: hidden;` was needed but even for
very long results, this does not seem to make any difference.

I noticed that it was already part of the initial SG open-source release
so there's likely no context around that anymore. 😅