experimental search input: Remove filter field background
We deciced to remove the background color and continue to use the same
styling we already have. By removing it we also magically fix this part
of the dark mode problem ;)

Whether or not to keep the context filter border is still up for debate
but I've slightly tweaked the border color for now to make it more
visible in dark mode.