Run VSCode build on main CI again
This brings back some stuff from the VS Code pipeline into `main`.

We're now at least building the VSCode extension when `client/` files
change. This is necessary because we were breaking the build setup too
often which took a while to fix and made it harder for us to push
fixes.

The E2E tests are still disabled until we investigate into the
flakiness.