permissions-center: remove unnecessary `useEffect` usage for jobs refetch.
The reason that `useEffect` is not needed is that when the filter is selected, component (`PermissionsSyncJobsTable`) state is changed at it is being re-rendered with new parameters, hence new GraphQL query is sent.

Also, `refetch` call in `useEffect` was never made with the same reason -- filters and GraphQL query variables were never different because of component state update.

Test plan:
Local sg run and manual tests.