graphqlbackend: add APIs for local directories
This implements a proposal for GraphQL APIs to preview the repositories
under a directory on the machine Sourcegraph is running. This is
intended only to be run for Sourcegraph App, and as such will only run
for admins on that deploy type.

The shape of this API is likely to change. However, we would like to
land this since it is a building block for the wizard. In particular I
imagine we may scrap this API and instead have a generic "preview" API
for any external service.

Test Plan: TODO update. I intend to run some tests against the GraphQL
API once the power returns in ZA for me to use my desktop machine. Will
update with queries then.