sg: bazel command sets block until bazel build is complete before anything else. (#48620)
Early feedback we got was that when using the bazel commandsets, if the
user never ran bazel build before, the logs would spam errors because
normal services would come up way earlier and complained about not being
able to reach other services.

While is never happens if you have only a few things to build, the
initial build command on a cold environment is really long. This fixes
this. In essence, it behaves like the old install method for those
services.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Tested locally.