Treeview: Fix invalid tabIndex setting when loading from a blob view
There's a bug in the treeview component where the tabbable node can be
set to the wrong item when expanding is called with the root node
(id: 0).

This manifests in the fact that tabbing into the treeview does not work
anymore because none of the rendered items will have a `tabIndex={0}`
set because the treeview thinks that the root item should receive the
focus but it's not rendered.

We fix this in the Wildcard component by removing `0` from the list of
expanded IDs.