Simplify automating batch changes in-app changelog for release
We usually decide manually when to add it, then add entries and make sure it lands before the release. If we forget, there would be a weird old banner showing on newer instances, so we should at least make a call to hide it (which we currently do by commenting it out), or add items to it.

New process suggestion:
- Batchers make sure that they include their in-app changelog entry in the PR the feature was created in
- The release tooling does what I documented inline, which essentially replaces two variables after the cut. That should probably happen right after the branch is cut, not when the release is going out.