own: ingestion UI (#48642)
Closes https://github.com/sourcegraph/sourcegraph/issues/47976
Closes https://github.com/sourcegraph/sourcegraph/issues/47826

Allows uploading and deleting ingested CODEOWNERS files by admins.


![image](https://user-images.githubusercontent.com/206864/222804581-39215178-24dd-490f-bd90-bcfb63438867.png)

![image](https://user-images.githubusercontent.com/206864/222804603-5b9d0b8c-f071-4a40-a26a-a1e22d99ce75.png)


Allows any user to view the ingested CODEOWNERS file. Line numbers in
URL query
(`https://sourcegraph.test:3443/github.com/hashicorp/errwrap/-/own?L=500`)
are supported, the line will be highlighted and the page will scroll to
the line.


![image](https://user-images.githubusercontent.com/206864/222804113-f09a2dd0-a6ee-4f67-ac07-a25fbc56c568.png)

The file limit for ingested CODEOWNERS is 10MB. This is mostly
arbitrary.


![image](https://user-images.githubusercontent.com/206864/222804276-f25a9c5b-2926-4da5-8dab-464a8c14d8e3.png)

The page is accessible from the repo homepage:

![CleanShot 2023-03-03 at 10 58
48](https://user-images.githubusercontent.com/206864/222804387-f9182d98-5cfb-4e37-9a09-cb5a9e254337.png)

The ownership page is inaccessible if `search-ownership` feature flag is
not enabled. The button won't show in the repo homepage and the `/-/own`
URL will redirect to the repo homepage.

## Test plan

Added storybook tests for the matrix {admin, non-admin} x {empty,
populated}.