remove queryTransformer API and undocumented go.imports: search operator (#48686)
There was an undocumented `go.imports:` search operator that did a hacky
attempt to find Go files that imported a given package. This was never
documented and was the only known use of the query transformer API in
the deprecated Sourcegraph extension API. This removal is part of the
removal of the deprecated Sourcegraph extension API. The upcoming
support for packages in Sourcegraph will be a better way to support this
in the future.

## Test plan

The existing search integration tests are sufficient to ensure that this
does not break searches in general.