Fix a batch of eslints (#80)
31 errors and 348 warnings &rarr; 3 errors and 264 warnings.

The remaining errors are wandering into `resources` and `dist` which are
not TypeScript, we can worry about that later.

Improve extension debugging launch. Note you must be running a Cody
server on port 49300 and embedding server on 49301 for the tests to pass,
and change the src/test/suite/index.ts Mocha timeout to give you time to
debug interactively.