[feat] Prepare for permissions data model oob migration (#48709)
## Description

Created a new column `migrated` in `user_permissions` table. By default
it will be `TRUE`, but all the existing rows will be marked as `false`
This way, we know if the row was already migrated to new unified table
or not.

Changed `perms_syncer` to write to both unified `user_repo_permissions`
and old `user_permissions` tables at once. Perms syncer will write
`true` to the above column for all the rows that it does touch in
`user_permissions` table. This way, `perms_syncer` is already doing part
of the job for the OOB migration.

## Test plan

Tested that write works for both tables locally.

Tested that the migration SQL queries work as expected.