filepicker: introduce package for local file dialogs (#48712)
This package implements a best-effort approach to running a file dialog
selector. Instead of linking in a GUI framework into Sourcegraph App, we
rely on shelling out to common utilities used by shell scripts to run a
dialog.

For Mac we can use osascript which is always available. For Linux (and
unofficially BSD/etc) we can rely on what the display manager provides
(zenity for gnome/gtk environments, kdialog for KDE).

For now this commit only integrates it into the test program
app-discover-repos for testing. We intend to integrate this into the
Sourcegraph App though via an API.

Test Plan: app-discover-repos -picker on linux with and without zenity
and kdialog. The same on darwin, which always worked. Also manually
changed app-discover-repos to pass in a short context timeout, an error
was returned as expected.

Based on that manual testing I created unit tests which simulate what
the programs do.

Part of https://github.com/sourcegraph/sourcegraph/issues/48127