experimental search input: Fix repo and file value completions (#48730)
This was introduced in #48491. The suggestions explicitly insert
`repo:`/`file:` to the value and I didn't test the change properly.

This commit updates the completion functions to include the goto action
and switch back to using that.


## Test plan

- Type into query input, select repo suggestion -> suggestion gets
inserted with `repo:` prefix
- Type into query input, select file suggestion -> suggestion gets
inserted with `file:` prefix
- Type `repo:` into input -> suggestions include "go to" action
  - Applying a suggestion with "Enter" inserts the repo name only 
  - Applying a suggestion with "Shift-Enter" goes to repo
- Type `file:` into input -> suggestions include "go to" action
  - Applying a suggestion with "Enter" inserts the file name only 
  - Applying a suggestion with "Shift-Enter" goes to file

## App preview:

-
[Web](https://sg-web-fkling-search-input-double-filter.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.