insights: update backfill ping query to use the backfill table directly (#48137)
Uses the backfill and repo iterator tables directly to determine
duration of the backfill to be more accurate.

Sets the "all repos" value of the ping to always be false. With the new
repo selection option this is no longer useful, we could possible
replace with a median # of repos or something to help quantify the
durations.

Ping Fields
AllRepos - Always false as this is no longer relevant because users have
more options between some/all repos
Count - The number of insight series that have been backfilled during
the current week
P99Seconds - P99 seconds of active runtime duration 
P90Seconds - P90 seconds of active runtime duration 
P50Seconds - P50 seconds of active runtime duration 

Also removes the legacy background job for calculating the backfill
completed at time since this is now handled in the backfiller.