nogo: create config for all analyzers (#48711)
"_base" is a special config key is inherited by all analyzers, if those
that are not specified. A key thing to remember is that if a child
config specified it **will overwrite** the inherited config key. For
example, in the current `nogo` config, `bodyclose` overwrites the
`exclude_files` config. Thus, we repeat the base config exclusions*
## Test plan
* cherry-picked this commit onto
https://github.com/sourcegraph/sourcegraph/pull/48696
* `bazel test //...` - no nogo failures

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->