gitserver: Don't load entire PyPi zips into memory
I've looked at our profiler and noticed that we use a LOT of memory in this single function. Since unzipping requires a ReaderAt, the most efficient way is to buffer the zip to a temp file. The temp file directory is managed by gitserver so it
- lives inside the mounted disk, so we don't need to worry about write permissions
- is cleaned up by gitserver, so we don't need to worry about potential FS pollution

The tests still pass, so this should be good? Let me know how I could test this, as I don't know much about packages on local dev.