Repo syncing: ingest topics as metadata (#47635)
We currently support searching repos by their attached metadata, but
ingestion of metadata is currently manual, so the feature doesn't get a
lot of use. This updates our repo syncing code to pull topics from
GitHub during sync. There will be a followup PR that uses this for a
`repo:has.topic()` predicate.