graphqlbackend: remove parameter gore in NewSchema (#48755)
This change is motivated by the very large query params we pass to
NewSchema. Over a series of mechanical refactors we replace all the
optional resolvers with a struct OptionalResolver. We then use embedding
so everything just keeps ticking along like before.

The main downside I imagine of this change is it is slightly easier to
forget to set the field when we call it for reals in frontends setup.
But you could of just as easily forget to set the fields in the
enterpriseServices struct before.

Test Plan: CI