Remove some unused feature flags (#47667)
This removes _some_ of the feature flags mentioned in
https://github.com/sourcegraph/sourcegraph/issues/45048.

Namely:

- `experimentalFeatures.enablePostSignupFlow`
- `open-beta-enabled` (was already removed)
- `org-code` (was already removed)
- `org-deletion` (removes flag and dotcom-feature-flagged behaviour)
- `github-app-cloud` (that one was unused, so now it's removed)
- `props.authenticatedUser?.tags?.includes('CreateOrg')` (removed)
- `externalServiceUserMode` (removed the whole `cloud-ga.ts` file, since
it was unused)
- `AllowUserExternalServicePrivate` user tag (removed, because unused)
- `AllowUserExternalServicePublic` user tag (removed, because unused)

What this PR does **not** remove yet:

- `org-email-invites`

I haven't figured out the code completely yet. I think if the feature
flag is off we allow invites only by username. Only if flag is on do we
allow invite by email. But I'm not 100% sure yet.


## Test plan

- CI