remove saved searches functionality from the VS Code extension (#48705)
Saved searches is a minor Sourcegraph feature that is mostly superseded
by code monitoring. This was used by ~0.001% of VS Code users, according
to our telemetry. That low usage, combined with the need for us to
reduce the overlap between saved searches and code monitoring in the
future, motivates removing it.

## Test plan

n/a; removal only