extsvc: Other defines root path for git discovery (#47779)
to be merged into feature branch of PR
https://github.com/sourcegraph/sourcegraph/pull/47673/

In PR 47673 serve-git at startup identifies a relevant root directory to
be walked for local repo sync'ing. The external service configuration
created at startup will define the url and not much else. That ext svc
configuration does not define the directories, so we can't make updates
to which directories are relevant; that is, it's fixed to what's defined
by `SRC` env var.

This pull request allows external service kind Other to store the root
directories in the configuration. Also, we change how serve-git handles
list-repos in that it accepts a list of the directories to walk.

There are still open questions that I'll try to cover in the comments.

## Test plan

sg start app
unit tests

---------

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>