Search suggestions: add file path to symbol suggestions
Previously, when accepting a symbol suggestion we added `type:symbol` to
the query and nothing more. This approach has surprising behavior when
there are many symbol suggestions with the same name because it  doesn't
matter which of the suggestions the user accepts, they all add the same
`type:symbol` syntax, which produces the same search results. Now, we
also add the file path so that the search results prioritize the
accepted symbol even if there are other symbols with the same name that
are defined in other files.

Fixes https://sourcegraph.slack.com/archives/CHEKCRWKV/p1678202639864099