experimental search input: Fix filter value completion when filter is not at the end
For better or worse we are showing filter value completions for a filter
as long as the cursor is anywhere inside the filter. But if the filter
doesn't have a value then we are inserting the value at [filter.end,
current cursor postion] which only works if the cursor is at the end of
the cursor.

This change ensures that the value is always correctly inserted after
the field.