experimental serch input: Add "last search context" suggestion
We've gotten feedback that because the search context is not a separate
dropdown anymore it's easy to "loose" your current search context by
e.g. clearing out the query.

The changes here build on top of the refactor in #48924 and introduces a
new suggestion source for the "last used search context".

The extension will show a suggestion if the query is empty or does not
contain a `context:` filter and the cursor is at a whitespace token.

The suggested search context will either be 'globally' selected one or
the one that was previously present in the query. A state field is used
to keep track of the currently/most recently present search context
value.