Ranking: solidify ranking toggle
This simplifies the approach to the ranking toggle and fixes a few bugs. Now, we
only disable ranking if the search-ranking feature flag is set, and the toggle
has been explicitly disabled. Otherwise we don't touch anything and use the
standard backend behavior.

This fixes a bug where on page reload, we would not use search ranking even the
toggle is enabled. I think this happened because feature flags are loaded
asynchronously, so the feature flag was undefined which we interpreted as
'disable ranking'.