Ranking: solidify ranking toggle (#48965)
This simplifies the approach to the ranking toggle and fixes a few bugs.
Now, we only disable ranking if the `search-ranking` feature flag is
set, and the toggle has been explicitly disabled. Otherwise we don't
touch anything and use the standard backend behavior.

This fixes a bug where on page reload, the search would not use ranking
even the toggle was enabled. I think this happened because feature flags
are loaded asynchronously, so the feature flag was undefined which we
interpreted as 'disable ranking'.

## Test plan

Existing tests, a lot of manual testing.