Use uncached doer for downloading packages
We've encountered that all our external requests are cached in redis, which means (1) that we need to read the entire body into memory so we can serialize it for redis, and (2) that we store potentially gigantic binary files in redis, which we never need to fetch again. This introduces an uncached version of the external doer, adds comments on when to use which, and modifies each of the packages clients to use that doer for the Download operation.