App: Add access token callback page for App
Part of #47998

This adds a new allowed requestor to the user access token callback page
that will allow App to make a request to dotcom to "sign in" a user.

Under the hood this uses the same logic that we added for the VS Code
extension: The user is redirected to a page on Sourcegraph that will
create an access token automatically and redirect the user back to App.

🚨 SECURITY: This flow currently does not need a user to accept the
token. This means that if a user has malware installed on their local
machine that intercepts `vscode://` links or listens on `localhost:3080`
they can create an access token for the user without any user input.

Should we update the step to require the user to confirm the creation of
the token?

To make sure we do not cause any additional security issues, the new
App requestor is only allowed on dotcom.