packages: filter check resolver handles version paging distinctly (#49023)
Previously, when checking for package versions matching a filter, we
would reuse the package repo connection resolver, but with an element of
max 1. This was problematic as we still applied the user input limit to
the versions returned in that one package, but without a way to pass in
a cursor based on the version paging. This PR addresses that by making
the graphql endpoint for checking for matches for a filter return a
union of either a connection resolver on packages _or_ versions.


![image](https://user-images.githubusercontent.com/18282288/224071539-2ad36279-d837-4d8f-bc66-7f252130bdc7.png)


![image](https://user-images.githubusercontent.com/18282288/224071782-57489af7-d52e-489f-ac18-b842acee9915.png)

## Test plan

Tested in graphql API console locally, included screenshots as :b:roof