Revert forced NODE_ENV to fix app build
After merging #49019, we started seeing broken App builds because
apparently part of the compiled assets still think that they are in dev
mode and some parts are not.

I assume this meant that a previous version was running React in
development mode, which is something that we actually need to fix up in
the future.

For now, to unblock the soft-launch, we're reverting the change that
forced a NODE_ENV in the `package.json` and instead add a new env
variable to ensure that the fix still works.