SCIM: Keep user emails in sync with SCIM patch and replace requests (#48980)
This adds logic to ensure the user_emails table in updated in response
to the PATCH and PUT requests that modify emails.

closes https://github.com/sourcegraph/sourcegraph/issues/48720

Notes: 

- Currently the create endpoint does not enforce that all emails are
successfully created, it only enforces that "primary" emails are
created. We may want to reconsider this decision as it can make the SCIM
data and the emails table out of sync.
- The PATCH/PUT endpoints **do** enforce that all requested complete
successfully however they do also account to the possibility that emails
may exist that SCIM doesn't know about and will reuse these.
- The PATCH/PUT endpoints **will not** remove any additional emails that
were not added via SCIM. These remain in the table and can be
successfully added via future SCIM requests. Once SCIM is aware of the
emails they can be removed in future requests.
- SCIM allows multiple primary emails, however Sourcegraph does not in
the case of multiple primaries we take the first one.