Fix unable to select primary email when user has only one email (#48971)
Currently, when I create a user account that initially has no email
address stored, but then go in and add one, I still don't have a primary
email. Now, I would just need to go in and select it from the dropdown
(it actually _is_ already selected), but the dropdown and the Save
button are disabled. The only way to set a primary email address is to
add a second email, mark reload the page, now be able to select the
first one from the dropdown and hit Save, and then remove the second
bogus email address again.

This PR fixes that, by enabling the inputs if no primary email is
currently set, too. I've also added an empty element to the select in
that case, so that it correctly shows "empty" instead of the first email
in the list, which confused me and made me think "oh I _DO_ have a
primary email, why is this not showing correctly??". Blank will not be
an option once a primary email is selected once.

One question that I had while fixing this: Should we longer-term address
this by instead making sure that the first verified email is _always_
promoted to be the primary email as well? Kinda tedious if you have to
go in again and also set it as primary. Without having a primary email,
code monitors don't allow sending emails, for example.