experimental search input: Blur input on submission
Currently the search input stays focused when the query is submitted,
which implies showing suggestions. This is a problem on the serach
results page because it covers up part of the page.
While the input will be blurred automatically as soon as the results
arrived, the input itself shouldn't assume that it will be blurred via
an external event.

This PR updates the code to blur the input when the query is submitted.
It also fixes a bug where the input is refocused (and therefore the
suggestions are shown again) after the query is submitted when the
cursor is not positioned at the end of the input.
I wasn't able to find out _why_ this happens but I know it's related to
the fact that we update the selection in
`useUpdateEditorFromQueryState` (if I comment out the `selection` change
in line 416 the input is not focused again).
To avoid this behavior I simply avoid changing the selection when the
query hasn't changed and when the selection is not explicitly set.