packages: add env var for coursier cache and proper fallback to src_repos_dir (#48944)
If `COURSIER_CACHE_DIR` is set, it will be used directly. Else, if
`SRC_REPOS_DIR` is set, we will use `$SRC_REPOS_DIR/coursier`. If
neither are set, nothing will be created and gitserver will fail to
start
[anyways](https://sourcegraph.com/github.com/sourcegraph/sourcegraph@857fd9349995709af2b3d55540da180059db61f8/-/blob/cmd/gitserver/shared/shared.go?L96%3A11-98%3A3).

Additionally, this package is no longer imported into the `frontend` or
`repo-updater` binaries after commenting out an unused function.

Closes https://github.com/sourcegraph/sourcegraph/issues/48841

## Test plan

Ran locally to confirm the values of the env vars fallback properly