app: Add limited pings for Sourcegraph App (#48962)
Change the updatecheck pings for Sourcegraph App to use limited
telemetry.

This change introduces a limited ping request that contains only a small
subset of fields. The goal is to minimize the contents of the ping. The
fields that are retained in the limited ping are:

- `site` - the randomized anonymous site id
- `version` 
- `deployType` (a `single-program` value in this field identifies this
as a ping from Sourcegraph App)

 It adds the following new fields:

- `os` string: the operating system identifier ("mac", "linux",
"windows")
- `totalRepos` integer: total number of repositories added.
- `activeToday` boolean: whether the user has been active today, based
on the existing function `getUsersActiveTodayCount`

This change also adds `omitempty` on all the non-required fields in the
ping request. This cleans up the JSON that is sent between the
updatecheck client and the handler, omitting empty fields. The goal of
adding `omitempty` is to avoid any confusion about the contents of the
request, by excluding empty ping fields that aren't used in App anyway.

## Test plan

- Added test for limited ping request

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>
Co-authored-by: Stephen Gutekanst <stephen@sourcegraph.com>