[feat] migrate perms_store scheduling functions to use perms_sync_jobs_history (#49078)
## Related to

This is the same code as here, that was already approved, but merging to
main instead:
- https://github.com/sourcegraph/sourcegraph/pull/48915

## Description
This change changes the way we schedule permission sync jobs.

We will follow these rules now:

- For the high priority schedule for entities with no permissions: List
all users who do not have permissions or do not have a job in
perms_sync_jobs_history table. And similar for the orthogonal repo case.
- For the medium priority schedue for entities with oldest permissions:
Take x amount of user_ids from perms_sync_jobs_history with the oldest
updated_at timestamps. Similar for orthogonal repo case.

This way, we simplify the logic for scheduling.
A nice side effect is also, that the scheduling of oldest perms is now
completely decoupled from the permissions themselves.

## Test plan
Added/modified unit tests. Tested locally that it works.