SCIM: Avoid auth warnings for SCIM requests (#49021)
For SCIM's REST endpoint, we use custom auth. Our current validation isn't prepared to recognize it, and it fails because it has a different format. The SCIM package has its own validation and error handling that's compatible with the SCIM format, so for that endpoint, we can safely avoid using our normal validation.