[feat] refactor GrantPendingPermissions to use unified perms (#49010)
## Description

Refactor GrantPendingPermissions to use unified perms.

Along the way, needed to refactor the input parameter for the
`GrantPendingPermissions` function.
As a result, needed to refactor the consumer of it as well, to send it
new type.
This was needed, because we need to store the external account ID as
well in the database now.
Luckily, the consumer already has this information, so we can just pass
it through.

When testing, I also noticed that I messed up before. When sending it
more than 65k parameters (the param limit), the
original query to `upsertUserRepoPermissions` failed. I wrote a small
helper to split the slice into reasonable sized
chunks and now we do mulitple queries with these chunks, so that these
big queries work.

## Test plan

Heavily unit tested.