experimental search input: Fix filter value completion when cursor is not at the end (#48934)
For better or worse we are showing filter value completions for a filter
as long as the cursor is anywhere inside the filter. But if the filter
doesn't have a value then we are inserting the value at [filter.end,
current cursor position] which only works if the cursor is at the end of
the cursor.

This change ensures that the value is always correctly inserted after
the field.



https://user-images.githubusercontent.com/179026/223786914-e7667943-ab37-4f5d-a183-8ff93d07f735.mp4



## Test plan
Manual testing (see video)

## App preview:

-
[Web](https://sg-web-fkling-search-input-fix-filter.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.