experimental serch input: Add "last search context" suggestion (#48958)
We've gotten feedback that because the search context is not a separate
dropdown anymore it's easy to "loose" your current search context e.g.
by clearing out the query.

The changes here build on top of the refactor in #48924 and introduces a
new suggestion source for the "last used search context".

The extension will show a suggestion if the query is empty or does not
contain a `context:` filter and the cursor is at a whitespace token.

The suggested search context will either be 'globally' selected one or
the one that was previously present in the query. A state field is used
to keep track of the currently/most recently present search context
value.



https://user-images.githubusercontent.com/179026/223855084-1a032e81-c334-4147-89cf-b00d26811935.mp4



## Test plan

- Clear search query -> context suggestion should be the first
- Select a different search context then clear query -> the just
selected search context should be suggested

## App preview:

- [Web](https://sg-web-fkling-search-input-protect.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.