App: Add support to pass-through destination in app token creation flow
Part of #47998

For the App flow to "Connect your Sourcegraph account" we want to be
able to pass-through a destination parameters to the callback page. This
way we can redirect the user to the correct page after they have
generated a token.

We only support this for the App flow, and only when the destination
parameter starts with a `/` (to prevent open redirects).