packages: sync pkgrepo extsvc's after applying filters in background job (#49106)
Package repo references rely on repo-updater in order to add/remove
package repos as theyre listed from `lsif_dependency_repos`. To make
this happen sooner rather than later, we need to set the next_sync_at in
the external_services table. We could alternatively call a synchronous
endpoint on repo-updater directly, but this asynchronous version is good
enough.

## Test plan

Tested locally in a full-run of adding/updating a filter.