experimental search input: Show example query/filter suggestions
This PR adds a new suggestion source: examples. If the query is empty or
the user is at a whitespace token we are showing up to three examples at
the bottom to expose users to more advanced features.

When the user selects one of these examples (or explicitly hides it) we
mark it as "used" and don't show it again.

The examples are randomized every 10 minutes or whenever the list needs
to be updated to contain 3 items.