Don't keep own service cache warm forever in resolvers (#49110)
Currently, the own service has an internal cache. I want to move away
from that and split these two concerns up, but until then and as a short
term fix for starship, I've made it so that we recreate a new service
for a new resolver invocation, like we do for searches. This ofc means
that there's no caching between GitBlobOwnership calls, but it also
means that we're not hogging memory, or returning results that are a
month old because we don't invalidate caches.