Search: log more information about result click (#48970)
This PR adds two new fields to our click tracking telemetry:
* The number of results that were displayed. This lets us add filters in
our analysis like "median click position for searches with more than 5
results". This is helpful, because it's not a ranking success when there
is only one result and you click on it.
* Whether or not "intelligent ranking" was enabled. This lets us
separate out the click metrics for users with ranking enabled vs.
disabled.