fix: move musl config to .cargo/config.toml and now it works (#49131)
Previously, syntect_server ran and linked against C++ correctly in
"standard" build environments (read: non-musl envs). It also was able to
link (at least a compile time?) to C++ in musl environments. However(!)
When running this in the docker image, it was possible that
syntect_server would segfault when using a C++ grammar (often, it would
seg fault immediately).

This segfault is "uncatchable", not like the panics from `syntect` that
http-stablizer was designed to restart. So, this meant that whenever a
C++ grammar'd language would run, we'd lose a syntax highlighter
service. Well... when you refresh, you kill another one. Rinse and
repeat til even the other languages are affected.

Now, with the new build, we don't have that happen (you can verify this
by running the docker image locally and checking against C++ grammars)
and now syntax highlighting continues to work.

## Test plan

- [ ] Run this locally after rebuilding the docker image, and using the
docker image to supply highlighting (NOT your local cargo instance)
- [ ] Do it again, but now with your local cargo instance